+++ src/api-binder/index.ts
@@ -21,7 +21,7 @@ import log from '../lib/supervisor-console';
 
 import * as deviceState from '../device-state';
 import * as globalEventBus from '../event-bus';
-import * as TargetState from '../device-state/target-state';
+import * as TargetState from './poll';
 import * as logger from '../logger';
 
 import * as apiHelper from '../lib/api-helper';
